Rename message middleware
