+++ lib/middleware/user-inbound-message.js
